+++ src/main/java/org/opensearch/knn/index/query/KNNQueryFactory.java
@@ -129,6 +129,7 @@ public class KNNQueryFactory extends BaseQueryFactory {
         log.debug(String.format("Creating Lucene k-NN query for index: %s \"\", field: %s \"\", k: %d", indexName, fieldName, k));
         switch (vectorDataType) {
             case BYTE:
+            case BINARY:
                 return new LuceneEngineKnnVectorQuery(
                     getKnnByteVectorQuery(fieldName, byteVector, luceneK, filterQuery, parentFilter, expandNested)
                 );